Add lambda functions, lists and list indexation
Lambdas are defined on a single line using the syntax
`\(x,y,z) : x+y+z`. A `λ` can also be used, and parentheses are
optional.

Lists and list indexation are the same as in Python, with the difference
that the syntax `list[1,2,3]` returns a list with the values of
at indices.